requirements.txt
websocket-client
plotly
numpy
pandas
stockstats
stockplot
